Refactor to move suppliers to StandardAwsKmsClientSuppliers class
